+++ react/components/ProductSpecifications/Wrapper.js
@@ -41,4 +41,34 @@ const ProductSpecificationsWrapper = ({
   )
 }
 
+ProductSpecificationsWrapper.getSchema = () => {
+  return {
+      title: 'admin/editor.product-specifications.title',
+      description: '',
+      type: 'object',
+      properties: {
+        hiddenSpecifications: {
+          items: {
+            default: '',
+            type: 'string',
+            title: 'admin/editor.product-specifications.items.title'
+          },
+          description: 'admin/editor.product-specifications.hidden-specifications.description',
+          title: 'admin/editor.product-specifications.hidden-specifications.title',
+          type: 'array'
+        },
+        visibleSpecifications: {
+          items: {
+            default: '',
+            type: 'string',
+            title: 'admin/editor.product-specifications.items.title'
+          },
+          description: 'admin/editor.product-specifications.visible-specifications.description',
+          title: 'admin/editor.product-specifications.visible-specifications.title',
+          type: 'array'
+        }
+      }
+    }
+}
+
 export default ProductSpecificationsWrapper