+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Option to edit `visibleSpecifications` and `hiddenSpecifications` in Site Editor.
 
 ## [3.84.2] - 2019-11-12
 ### Fixed